+++ src/main/java/com/bookings/booking_management/service/Payment.java
@@ -1,5 +1,5 @@
 package com.bookings.booking_management.service;
 
 public interface Payment {
-    boolean processPayment(Long amount);
+    boolean processPayment(float amount);
 }
